+++ android/gradle.properties
@@ -1,5 +1,5 @@
-ReactNativeGeolocation_compileSdkVersion=31
-ReactNativeGeolocation_buildToolsVersion=28.0.3
-ReactNativeGeolocation_targetSdkVersion=31
-ReactNativeGeolocation_minSdkVersion=21
-NewBobGeo2_ndkversion=21.4.7075529
+RNCGeolocation_compileSdkVersion=31
+RNCGeolocation_buildToolsVersion=28.0.3
+RNCGeolocation_targetSdkVersion=31
+RNCGeolocation_minSdkVersion=21
+RNCGeolocation_ndkversion=21.4.7075529
