+++ example/android/settings.gradle
@@ -9,6 +9,3 @@ if (settings.hasProperty("newArchEnabled") && settings.newArchEnabled == "true")
     include(":ReactAndroid:hermes-engine")
     project(":ReactAndroid:hermes-engine").projectDir = file('../node_modules/react-native/ReactAndroid/hermes-engine')
 }
-
-include ':react-native-geolocation'
-project(':react-native-geolocation').projectDir = new File(rootProject.projectDir, '../../android')
